refact : API url 수정 및 로직 위치 변경
