+++ offroad-api/src/main/java/site/offload/api/adventure/dto/request/AuthPositionRequest.java
@@ -1,4 +1,4 @@
-package site.offload.api.member.dto.request;
+package site.offload.api.adventure.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
