+++ offroad-api/src/main/java/site/offload/api/member/controller/SocialLoginControllerSwagger.java
@@ -4,10 +4,13 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RequestBody;
+import site.offload.api.member.dto.request.LogoutRequest;
 import site.offload.api.member.dto.response.SocialLoginResponse;
+import site.offload.api.response.APIErrorResponse;
 import site.offload.api.response.APISuccessResponse;
 import site.offload.external.oauth.dto.SocialLoginRequest;
 
@@ -19,4 +22,11 @@ public interface SocialLoginControllerSwagger {
     ResponseEntity<APISuccessResponse<SocialLoginResponse>> login(
             @RequestBody SocialLoginRequest socialLoginRequest
     );
+
+    @Operation(summary = "로그아웃 API", description = "로그아웃 API")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "204", description = "로그아웃 성공"),
+            @ApiResponse(responseCode = "400", description = "로그아웃 실패", content = @Content(schema = @Schema(implementation = APIErrorResponse.class)))
+    })
+    ResponseEntity<APISuccessResponse<Void>> logout(@RequestBody final LogoutRequest request);
 }
