+++ offroad-api/src/main/java/site/offload/api/member/dto/request/LogoutRequest.java
@@ -2,7 +2,7 @@ package site.offload.api.member.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 
-public record SignOutRequest(
+public record LogoutRequest(
     @Schema(description = "Refresh Token", example = "refreshToken")
     String refreshToken
 ) {
