+++ offroad-api/src/main/java/site/offload/api/member/usecase/LogoutUseCase.java
@@ -3,16 +3,16 @@ package site.offload.api.member.usecase;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import site.offload.api.member.dto.request.SignOutRequest;
+import site.offload.api.member.dto.request.LogoutRequest;
 import site.offload.cache.config.RedisRepository;
 
 @Service
 @RequiredArgsConstructor
-public class SignOutUseCase {
+public class LogoutUseCase {
 
     private final RedisRepository redisRepository;
 
-    public void execute(SignOutRequest request) {
+    public void execute(LogoutRequest request) {
         redisRepository.delete(request.refreshToken());
     }
 }
