+++ offroad-api/src/test/java/site/offload/api/member/usecase/AuthAdventureUseCaseTest.java
@@ -6,6 +6,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import site.offload.api.adventure.usecase.AuthAdventureUseCase;
 import site.offload.api.character.service.CharacterService;
 import site.offload.api.charactermotion.service.CharacterMotionService;
 import site.offload.api.charactermotion.service.GainedCharacterMotionService;
@@ -13,8 +14,8 @@ import site.offload.api.coupon.service.CouponService;
 import site.offload.api.coupon.service.GainedCouponService;
 import site.offload.api.emblem.service.GainedEmblemService;
 import site.offload.api.exception.BadRequestException;
-import site.offload.api.member.dto.request.AuthAdventureRequest;
-import site.offload.api.member.dto.response.VerifyQrcodeResponse;
+import site.offload.api.adventure.dto.request.AuthAdventureRequest;
+import site.offload.api.adventure.dto.response.VerifyQrcodeResponse;
 import site.offload.api.member.service.MemberService;
 import site.offload.api.place.service.PlaceService;
 import site.offload.api.place.service.VisitedPlaceService;
